Fixup PPI by refreshing the tree on replacement
Another fix for GH #6

When removing elements in a PPI tree,
we are going to reach some unexpected state
and cannot access next elements.

This is an ugly fix, but this is better to refresh
the whole tree after performing a 'ref' update...
so we can iterate on a clean PPI tree.