rewrite_file overwrite previous content
Fixes GH #9

Pod mention that 'rewrite_file' "rewrite the file in place"
but obsiously it does not and simply return the updated content
instead.

This commit is adding a unit test and fixes 'rewrite_file' by
preserving the original behavior of returning the updated content.
(instead of a boolean as we could expect)